IOW-366 (Support 400, 404, and 500 by OGC guidance)   Swagger document updates:     Made the bbox parameter in endpoint a String     Updated BBox validator to take the Bounding Box single string value as       argument.     Restored ANC collection link to featureCollection_bbox_one.json     Added @parameter to add a description for bbox parameter in swagger ui     Moved openapi class to openapi.schema package
