+++ src/main/java/gov/usgs/wma/waterdata/collections/CollectionsController.java
@@ -15,15 +15,18 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-import gov.usgs.wma.waterdata.collections.geojson.CollectionGeoJSON;
-import gov.usgs.wma.waterdata.collections.geojson.CollectionsGeoJSON;
-import gov.usgs.wma.waterdata.collections.geojson.FeatureGeoJSON;
+import gov.usgs.wma.waterdata.openapi.schema.collections.CollectionGeoJSON;
+import gov.usgs.wma.waterdata.openapi.schema.collections.CollectionsGeoJSON;
+import gov.usgs.wma.waterdata.openapi.schema.collections.FeatureGeoJSON;
 import gov.usgs.wma.waterdata.parameter.BoundingBox;
 import gov.usgs.wma.waterdata.validation.BBox;
 
 import io.swagger.v3.oas.annotations.ExternalDocumentation;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.enums.ParameterIn;
 import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.ExampleObject;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -38,6 +41,8 @@ public class CollectionsController extends BaseController {
 
 	protected static final String LIMIT_VALIDATE_MESS = "limit must be greater than or equal to 1";
 	protected static final String START_INDEX_VALIDATE_MESS = "startIndex must be greater than or equal to 0";
+	protected static final String BBOX_DESCRIPTION = "Bounding box: minimum longitude, minimum latitude, maximum longitude, maximum latitude<br>"
+			+ "bbox=-109.046667,37.0,-102.046667,39.0 limits results to monitoring sites in Colorado.";
 
 	@Autowired
 	public CollectionsController(CollectionsDao collectionsDao, CollectionParams collectionsParams) {
@@ -103,14 +108,16 @@ public class CollectionsController extends BaseController {
 			},
 			externalDocs=@ExternalDocumentation(url="http://docs.opengeospatial.org/is/17-069r3/17-069r3.html#_feature_")
 		)
+	@Parameter(name="bbox", in=ParameterIn.QUERY, description=BBOX_DESCRIPTION)
 	@GetMapping(value = "collections/{collectionId}/items", produces = MediaType.APPLICATION_JSON_VALUE)
 	public String getOgcCollectionFeatures(
 			@RequestParam(value = "f", required = false, defaultValue = "json") String mimeType,
 			@Min(value=1, message = LIMIT_VALIDATE_MESS) @RequestParam(value = "limit", required = false, defaultValue = "10000") int limit,
 			@Min(value=0, message = START_INDEX_VALIDATE_MESS) @RequestParam(value = "startIndex", required = false, defaultValue = "0") int startIndex,
-			@BBox @RequestParam(value = "bbox", required = false) BoundingBox bbox,
+			@BBox @RequestParam(value = "bbox", required = false) String bbox,
 			@PathVariable(value = PARAM_COLLECTION_ID) String collectionId, HttpServletResponse response) {
 
+		BoundingBox boundingBox = bbox == null ? null : new BoundingBox(bbox);
 		int count = collectionsDao.getCollectionFeatureCount(collectionsParams.buildParams(collectionId));
 
 		String rtn;
@@ -119,7 +126,7 @@ public class CollectionsController extends BaseController {
 			rtn = ogc404Payload;
 		} else {
 			rtn = collectionsDao.getCollectionFeaturesJson(
-					collectionsParams.buildParams(collectionId, limit, startIndex, bbox, count));
+					collectionsParams.buildParams(collectionId, limit, startIndex, boundingBox, count));
 			if (rtn == null) {
 				response.setStatus(HttpStatus.NOT_FOUND.value());
 				rtn = ogc404Payload;
