+++ src/main/java/gov/usgs/wma/waterdata/location/MonitoringLocationController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 
-import gov.usgs.wma.waterdata.location.geojson.MonitoringLocationGeoJSON;
+import gov.usgs.wma.waterdata.openapi.schema.location.MonitoringLocationGeoJSON;
 import io.swagger.v3.oas.annotations.ExternalDocumentation;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
