+++ src/main/java/gov/usgs/wma/waterdata/openapi/schema/collections/FeatureGeoJSON.java
@@ -1,14 +1,13 @@
-package gov.usgs.wma.waterdata.collections.observations;
+package gov.usgs.wma.waterdata.openapi.schema.collections;
 
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
-import gov.usgs.wma.waterdata.geojson.Geometry;
+import gov.usgs.wma.waterdata.openapi.schema.geojson.Geometry;
 import io.swagger.v3.oas.annotations.media.Schema;
 
-
-@JsonPropertyOrder(value={"type","id","geometry","properties","timeSeries"})
-@Schema(description="Statistical Time Series Observation datasets available for a feature.")
-public class StatisticalFeatureGeoJSON  {
+@JsonPropertyOrder(value = { "type", "id", "geometry", "properties"})
+@Schema(description = "Feature MetaData in GeoJSON format.")
+public class FeatureGeoJSON {
 	@Schema(description="The type of this object - Currently only 'Feature' objects are produced.", example="Feature")
 	public String getType() {
 		return null;
@@ -22,12 +21,8 @@ public class StatisticalFeatureGeoJSON  {
 		return null;
 	}
 	@Schema(description="The properties associated with this Feature.")
-	public StatisticalFeatureProperties getProperties() {
+	public FeatureProperties getProperties() {
 		return null;
 	}
-	@Schema(description="The statistical time series available for this Feature.")
-	public StatisticalTimeSeriesItem[] getTimesSeries() {
-		return null;
-	}
-
+	
 }
