+++ src/main/java/gov/usgs/wma/waterdata/openapi/schema/location/MonitoringLocationProperties.java
@@ -1,8 +1,8 @@
-package gov.usgs.wma.waterdata.location.geojson;
+package gov.usgs.wma.waterdata.openapi.schema.location;
 
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
-import gov.usgs.wma.waterdata.geojson.Properties;
+import gov.usgs.wma.waterdata.openapi.schema.geojson.Properties;
 
 @JsonPropertyOrder(value={"samplingFeatureName"})
 public class MonitoringLocationProperties extends Properties {
