+++ src/main/java/gov/usgs/wma/waterdata/openapi/schema/observations/ObservationsJSON.java
@@ -1,4 +1,4 @@
-package gov.usgs.wma.waterdata.collections.observations;
+package gov.usgs.wma.waterdata.openapi.schema.observations;
 
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
