+++ src/main/java/gov/usgs/wma/waterdata/validation/BBoxValidator.java
@@ -7,7 +7,7 @@ import javax.validation.ConstraintValidatorContext;
 
 import gov.usgs.wma.waterdata.parameter.BoundingBox;
 
-public class BBoxValidator implements ConstraintValidator<BBox, BoundingBox> {
+public class BBoxValidator implements ConstraintValidator<BBox, String> {
 
 	public static final BigDecimal MAX_LATITUDE = new BigDecimal(90);
 	public static final BigDecimal MIN_LATITUDE = new BigDecimal(-90);
@@ -15,14 +15,15 @@ public class BBoxValidator implements ConstraintValidator<BBox, BoundingBox> {
 	public static final BigDecimal MIN_LONGITUDE = new BigDecimal(-180);
 
 	@Override
-	public boolean isValid(BoundingBox value, ConstraintValidatorContext context) {
+	public boolean isValid(String bboxSingleValue, ConstraintValidatorContext context) {
 		boolean valid = false;
-		if (null == value) {
+		if (null == bboxSingleValue) {
 			valid = true;
-		} else if (value.getSingle() == null || value.getSingle().split(",").length != 4) {
+		} else if (bboxSingleValue.split(",").length != 4) {
 			valid = false;
 		} else {
 			try {
+				BoundingBox value = new BoundingBox(bboxSingleValue);
 				BigDecimal north = new BigDecimal(value.getNorth());
 				BigDecimal south = new BigDecimal(value.getSouth());
 				BigDecimal east = new BigDecimal(value.getEast());
